Seprate profile utilities to `config_util.go`.
